Bugfixes, cleanup and refactoring
