+++ src/Components/Options.jsx
@@ -1,4 +1,9 @@
-import React, { memo, useCallback } from 'react';
+import React, {
+    memo,
+    useCallback,
+    useEffect,
+    useRef,
+} from 'react';
 import PropTypes from 'prop-types';
 import Option from './Option';
 import isSelected from '../lib/isSelected';
@@ -13,6 +18,7 @@ const Options = ({
     renderGroupHeader,
     snapshot,
 }) => {
+    const selectRef = useRef(null);
     const renderEmptyMessage = useCallback(() => {
         if (emptyMessage === null) {
             return null;
@@ -22,42 +28,63 @@ const Options = ({
 
         return <li className={cls('not-found')}>{content}</li>;
     }, [emptyMessage, cls]);
+    const { focus, value, highlighted } = snapshot;
+
+    useEffect(() => {
+        const { current } = selectRef;
+
+        if (!current || Array.isArray(value) || (highlighted < 0 && value === undefined)) {
+            return;
+        }
+
+        const query = (highlighted > -1) ? `[data-index="${highlighted}"]` : `[data-value="${escape(value)}"]`;
+        const selected = current.querySelector(query);
+
+        if (selected) {
+            const rect = current.getBoundingClientRect();
+            const selectedRect = selected.getBoundingClientRect();
+
+            current.scrollTop = selected.offsetTop - (rect.height / 2) + (selectedRect.height / 2);
+        }
+    }, [focus, value, highlighted, selectRef]);
 
     return (
-        // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions
-        <ul className={cls('options')}>
-            {options.length > 0 ? (
-                options.map((option) => {
-                    const isGroup = option.type === 'group';
-                    const items = (isGroup) ? option.items : [option];
-                    const base = { cls, optionProps, renderOption };
-                    const rendered = items.map((o) => (
-                        <Option
-                            key={o.value}
-                            selected={isSelected(o, snapshot.option)}
-                            highlighted={snapshot.highlighted === o.index}
-                            {...base}
-                            {...o}
-                        />
-                    ));
+        // eslint-disable-next-line jsx-a11y/no-static-element-interactions
+        <div className={cls('select')} ref={selectRef} onMouseDown={(e) => e.preventDefault()}>
+            <ul className={cls('options')}>
+                {options.length > 0 ? (
+                    options.map((option) => {
+                        const isGroup = option.type === 'group';
+                        const items = (isGroup) ? option.items : [option];
+                        const base = { cls, optionProps, renderOption };
+                        const rendered = items.map((o) => (
+                            <Option
+                                key={o.value}
+                                selected={isSelected(o, snapshot.option)}
+                                highlighted={snapshot.highlighted === o.index}
+                                {...base}
+                                {...o}
+                            />
+                        ));
 
-                    if (isGroup) {
-                        return (
-                            <li role="none" className={cls('row')} key={option.groupId}>
-                                <div className={cls('group')}>
-                                    <div className={cls('group-header')}>{renderGroupHeader(option.name)}</div>
-                                    <ul className={cls('options')}>
-                                        {rendered}
-                                    </ul>
-                                </div>
-                            </li>
-                        );
-                    }
+                        if (isGroup) {
+                            return (
+                                <li role="none" className={cls('row')} key={option.groupId}>
+                                    <div className={cls('group')}>
+                                        <div className={cls('group-header')}>{renderGroupHeader(option.name)}</div>
+                                        <ul className={cls('options')}>
+                                            {rendered}
+                                        </ul>
+                                    </div>
+                                </li>
+                            );
+                        }
 
-                    return rendered;
-                })
-            ) : (renderEmptyMessage() || null)}
-        </ul>
+                        return rendered;
+                    })
+                ) : (renderEmptyMessage() || null)}
+            </ul>
+        </div>
     );
 };
 
@@ -81,6 +108,12 @@ Options.propTypes = {
     snapshot: PropTypes.shape({
         highlighted: PropTypes.number.isRequired,
         option: PropTypes.oneOfType([optionType, PropTypes.arrayOf(optionType)]),
+        focus: PropTypes.bool.isRequired,
+        value: PropTypes.oneOfType([
+            PropTypes.string,
+            PropTypes.number,
+            PropTypes.arrayOf(PropTypes.oneOfType([PropTypes.string, PropTypes.number])),
+        ]),
     }).isRequired,
     renderOption: PropTypes.func,
     renderGroupHeader: PropTypes.func,
