+++ src/types.js
@@ -16,13 +16,3 @@ export const optionType = PropTypes.oneOfType([
         items: PropTypes.arrayOf(option),
     }),
 ]);
-
-export const valueType = PropTypes.oneOfType([
-    PropTypes.object,
-    PropTypes.arrayOf(PropTypes.object),
-]);
-
-export const classNameType = PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.func,
-]);
